backend/requirements.txt
django==5.0.5
djangorestframework==3.15.1
pillow==10.2.0
mysql-client==0.0.1
django-cors-headers==4.3.1
django-jazzmin==2.6.1